+++ frontend/src/app/components/containers/complaints/allegations/allegation-complaint-list-item.tsx
@@ -1,27 +1,27 @@
 import { FC } from "react";
-import { HwcrComplaint } from "../../../types/complaints/hwcr-complaint";
-import { formatDateTime } from "../../../common/methods";
-import ComplaintEllipsisPopover from "./complaint-ellipsis-popover";
+import ComplaintEllipsisPopover from "../complaint-ellipsis-popover";
+import { AllegationComplaint } from "../../../../types/complaints/allegation-complaint";
+import { formatDateTime } from "../../../../common/methods";
 
 type Props = {
   type: string;
-  complaint: HwcrComplaint;
+  complaint: AllegationComplaint;
   complaintClick: Function;
   sortKey: string;
   sortDirection: string;
 };
 
-export const WildlifeComplaintListItem: FC<Props> = ({
+export const AllegationComplaintListItem: FC<Props> = ({
   type,
   complaint,
   complaintClick,
-  sortKey, 
-  sortDirection
+  sortKey,
+  sortDirection,
 }) => {
   const {
     complaint_identifier: complaintIdentifier,
-    hwcr_complaint_nature_code,
-    species_code,
+    violation_code,
+    in_progress_ind,
   } = complaint;
 
   const {
@@ -35,20 +35,18 @@ export const WildlifeComplaintListItem: FC<Props> = ({
   } = complaintIdentifier;
 
   const incidentReportedDatetime = formatDateTime(incident_reported_datetime);
-  const natureCode =
-    hwcr_complaint_nature_code !== null
-      ? hwcr_complaint_nature_code.long_description
-      : null;
-  const species = species_code.short_description;
+
   const location = cos_geo_org_unit ? cos_geo_org_unit.area_name : null;
 
   const firstName = person_complaint_xref[0]?.person_guid?.first_name;
   const lastName = person_complaint_xref[0]?.person_guid?.last_name;
+
   const firstInitial = firstName?.length > 0 ? firstName.substring(0, 1) : "";
   const lastInitial = lastName?.length > 0 ? lastName.substring(0, 1) : "";
   const initials = firstInitial + lastInitial;
+
   const displayName =
-    firstInitial.length > 0 ? firstInitial + ". " + lastName : lastName;
+    firstInitial.length > 0 ? `${firstInitial}. ${lastName}` : lastName;
 
   const statusButtonClass =
     complaint_status_code.long_description === "Closed"
@@ -56,16 +54,22 @@ export const WildlifeComplaintListItem: FC<Props> = ({
       : "btn btn-primary comp-status-open-btn";
   const status = complaint_status_code.long_description;
 
-  const updateDate =
-    Date.parse(update_timestamp) >= Date.parse(update_timestamp)
-      ? formatDateTime(update_timestamp)
-      : formatDateTime(update_timestamp);
+  const updateDate = formatDateTime(update_timestamp);
 
-  const zone = cos_geo_org_unit.zone_code;
-  const assigned_ind = person_complaint_xref.length > 0 && person_complaint_xref[0].active_ind;
+  const assigned_ind =
+    person_complaint_xref.length > 0 && person_complaint_xref[0].active_ind;
 
-    return (
-    <tr key={id}>
+  const violationCode =
+    violation_code != null ? violation_code.long_description : "";
+  const inProgressButtonClass =
+    String(in_progress_ind) === "true"
+      ? "btn btn-primary comp-in-progress-btn"
+      : "btn btn-primary comp-in-progress-btn btn-hidden";
+
+  const inProgressInd = String(in_progress_ind) === "true" ? "In Progress" : "";
+
+  return (
+    <tr>
       <td
         className="comp-cell-width-95 comp-header-left-border"
         onClick={(event) => complaintClick(event, id)}
@@ -73,27 +77,22 @@ export const WildlifeComplaintListItem: FC<Props> = ({
         {id}
       </td>
       <td
-        className="comp-cell-width-95 comp-header-vertical-border"
+        className="sortableHeader comp-cell-width-95 comp-header-vertical-border"
         onClick={(event) => complaintClick(event, id)}
       >
         {incidentReportedDatetime}
       </td>
+      <td className="sortableHeader comp-cell-width-305">{violationCode}</td>
       <td
-        className="comp-cell-width-330"
+        className="sortableHeader comp-cell-width-155 comp-header-vertical-border"
         onClick={(event) => complaintClick(event, id)}
       >
-        {natureCode}
-      </td>
-      <td
-        className="comp-cell-width-130 comp-header-vertical-border"
-        onClick={(event) => complaintClick(event, id)}
-      >
-        <button type="button" className="btn btn-primary comp-species-btn">
-          {species}
+        <button type="button" className={inProgressButtonClass}>
+          {inProgressInd}
         </button>
       </td>
       <td
-        className="comp-cell-width-165"
+        className="sortableHeader comp-cell-width-165"
         onClick={(event) => complaintClick(event, id)}
       >
         {location}
@@ -105,17 +104,17 @@ export const WildlifeComplaintListItem: FC<Props> = ({
         {locationSummary}
       </td>
       <td
-        className="comp-cell-width-130"
+        className="sortableHeader comp-cell-width-130"
         onClick={(event) => complaintClick(event, id)}
       >
         <div
           data-initials-listview={initials}
           className="comp-profile-avatar"
-        ></div>{" "}
+        ></div>
         {displayName}
       </td>
       <td
-        className="comp-cell-width-75 comp-header-vertical-border"
+        className="sortableHeader comp-cell-width-75 comp-header-vertical-border"
         onClick={(event) => complaintClick(event, id)}
       >
         <button type="button" className={statusButtonClass}>
@@ -123,7 +122,7 @@ export const WildlifeComplaintListItem: FC<Props> = ({
         </button>
       </td>
       <td
-        className="comp-cell-width-110 comp-header-right-border"
+        className="sortableHeader comp-cell-width-110 comp-header-right-border"
         onClick={(event) => complaintClick(event, id)}
       >
         {updateDate}
@@ -132,10 +131,9 @@ export const WildlifeComplaintListItem: FC<Props> = ({
         complaint_identifier={id}
         complaint_type={type}
         assigned_ind={assigned_ind}
-
         sortColumn={sortKey}
         sortOrder={sortDirection}
-      ></ComplaintEllipsisPopover>
+      />
     </tr>
   );
 };
